Moved qtclient to a different directory.
