+++ qspade/editorwidget.py
