+++ qspade/mainwindow.py
